pyproject.toml
[tool.poetry]
name = "langchain-files"
version = "0.1.0"
description = "AI agent for file system interaction"
authors = ["Nirajan Sangraula"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
python-dotenv = "^1.1.1"
langchain = "^0.3.26"
langchain-openai = "^0.3.27"
duckdb = "^1.3.1"
langchain-community = "^0.3.27"
duckdb-engine = "^0.17.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
